refactor user top routes
